refactor(client): factorize screen to suggest tracks
