+++ packages/client/screens/MusicPlaylistEditorSearchTracksScreen/index.tsx
@@ -1,18 +1,7 @@
-import { useActor, useMachine } from '@xstate/react';
-import { View } from 'dripsy';
-import { FlatList } from 'react-native';
-import React, { useState } from 'react';
-import { ActorRef } from 'xstate';
-import TrackListItem from '../../components/Track/TrackListItem';
-import { AppScreenWithSearchBar } from '../../components/kit';
-import {
-    AppScreenHeaderWithSearchBarMachineEvent,
-    AppScreenHeaderWithSearchBarMachineState,
-} from '../../machines/appScreenHeaderWithSearchBarMachine';
-import { searchTrackMachine } from '../../machines/searchTrackMachine';
-import { assertEventType } from '../../machines/utils';
+import React from 'react';
 import { MpeTabMpeSearchTracksScreenProps } from '../../types';
 import { usePlaylist } from '../../hooks/useMusicPlaylistsActor';
+import AppSuggestTrackScreen from '../../components/AppSuggestTrackScreen';
 
 const MusicPlaylistEditorSearchTracksScreen: React.FC<MpeTabMpeSearchTracksScreenProps> =
     ({
@@ -23,29 +12,15 @@ const MusicPlaylistEditorSearchTracksScreen: React.FC<MpeTabMpeSearchTracksScree
     }) => {
         const playlist = usePlaylist(playlistID);
         const playlistActorRef = playlist.ref;
-        const [screenOffsetY, setScreenOffsetY] = useState(0);
-        const [state, sendToSearchTracks] = useMachine(searchTrackMachine, {
-            actions: {
-                handleTrackPressed: (_, event) => {
-                    assertEventType(event, 'PRESS_TRACK');
-                    const { trackID } = event;
 
-                    playlistActorRef.send({
-                        type: 'ADD_TRACK',
-                        trackID,
-                    });
+        function handleTracksSelected([trackID]: string[]) {
+            playlistActorRef.send({
+                type: 'ADD_TRACK',
+                trackID,
+            });
 
-                    exitModal();
-                },
-            },
-        });
-        const tracksResults = state.context.tracks;
-        const searchBarActor: ActorRef<
-            AppScreenHeaderWithSearchBarMachineEvent,
-            AppScreenHeaderWithSearchBarMachineState
-        > = state.children.searchBarMachine;
-        const [searchState, sendToSearch] = useActor(searchBarActor);
-        const showHeader = searchState.hasTag('showHeaderTitle');
+            exitModal();
+        }
 
         function exitModal() {
             navigation.goBack();
@@ -55,50 +30,12 @@ const MusicPlaylistEditorSearchTracksScreen: React.FC<MpeTabMpeSearchTracksScree
             navigation.goBack();
         }
 
-        function handleTrackPress(trackID: string) {
-            return () => {
-                sendToSearchTracks({
-                    type: 'PRESS_TRACK',
-                    trackID,
-                });
-            };
-        }
-
         return (
-            <AppScreenWithSearchBar
-                canGoBack
-                title="Add Track"
-                searchInputPlaceholder="Search a track..."
-                showHeader={showHeader}
-                screenOffsetY={showHeader === true ? 0 : screenOffsetY}
-                setScreenOffsetY={setScreenOffsetY}
-                searchQuery={searchState.context.searchQuery}
-                sendToSearch={sendToSearch}
-                goBack={handleGoBack}
-            >
-                <FlatList
-                    data={tracksResults ?? []}
-                    renderItem={({
-                        item: { id, title, artistName },
-                        index,
-                    }) => (
-                        <View
-                            sx={{
-                                marginBottom: 'm',
-                            }}
-                        >
-                            <TrackListItem
-                                index={index + 1}
-                                title={title}
-                                trackID={id}
-                                artistName={artistName}
-                                onPress={handleTrackPress(id)}
-                            />
-                        </View>
-                    )}
-                    keyExtractor={(_, index) => String(index)}
-                />
-            </AppScreenWithSearchBar>
+            <AppSuggestTrackScreen
+                screenTitle="Add track"
+                onTracksSelected={handleTracksSelected}
+                onGoBack={handleGoBack}
+            />
         );
     };
